backend/requirements.txt
setuptools>=65.0.0
django==5.0.1
djangorestframework==3.14.0
pandas==2.1.4
openpyxl==3.1.2
psycopg2-binary==2.9.9
drf-yasg==1.21.7
django-cors-headers==4.3.1
